requirements.txt
fastapi
uvicorn
scikit-learn
nltk
ir_datasets
pandas
pyspellchecker
